Change FillArray API to BuildArray
FillArray expected a pre-constructed dense array. This was problematic in
implementation as FillArray had to ensure that it was initialized with holes
up to the specified length so that the kernel may run without bailing on
bounds checks. However, if we *did* bail, this resulted in a dense array with
JS_ARRAY_HOLE between its length and initializedLength being passed to the
self-hosted constructor. This violates the initializedLength invariant.

NB that we violate this invariant anyways for inside parallel code. But since
the write to the buffer is already unsafe, we just don't emit any
LStoreElementTs and always emit LStoreElementVs in Ion.