Removed isCall mark from ParLambda.
Like ParNew, enabled use of save/restoreLive of spilled registers by
initializing safepoint structure.  We may want to review whether the
full safepoint structure is overkill for our needs.

Note also that scopeReg is distinct from the live registers recorded
in the safepoint, and so the out-of-line code continues to stash it
via push/pop.  That distinction may warrant also further review; we
might be better trying to add it to the live registers in some way.