Defer triggering GC until the end of a fork-join section end.
Fixes the alloc-too-many-objs.js test, and generally seems to
be a relatively clean modification to the GC code.

As part of this patch, modify ParallelArray.js so that if a par mode is
specified, we do not try to run sequential.  This is very helpful for tests
that misbehave like allocating a lot of objects, because it means that we can
just test the behavior in par. mode and not have to wait for seq. mode to churn
through.