Fix bailout on map-nested
The bug is interesting. What happens is that somewhere along the line
heuristics were changed and we were hitting OSR'd Ion code where Ion didn't
kick in before. In the compiled loop, we inlined NewParallelArray into a call
to the custom init function, ParallelArrayView. This function returns void, so
the InvokeFunction VMFunction recorded void in the typeset of the call.