Fix another bug: hasIonScript needs to check for ION_COMPILING_SCRIPT.
This was causing Firefox to crash in the PJS builds.

The bug was injected by 8ca2c7647a0d467d9b7d45149108d06733448555 i.e.
Niko's merge with mozilla/master.  I do not yet know why git merged
the code in this broken manner, but I take it as a warning sign.