Added end-to-end RESIZ[E] benchmark, with seq and par variants.
Switch to using buildA instead of build in hopes that it speeds things
up.  (Have not verified this.)

To support the RESIZ benchmark:

  - Add computeEnergy method to RectArray to parallel the one on
    ParallelArray.

  - Split the cut seam methods into seq and par variants.

The RESIZ benchmark is very very slow at the moment.  I added
print-outs of intermediate state while it runs.  There is much to be
done to bring it in parity with Firefox in terms of exercising the js
shell in the same way.

(It also may currently expose crashiness in the PJS js shell build,
but I am not sure of this since I need to pull in changes from shu and
test them to see about that.)