Fixes to shell liquid-resize.js
My changes to use a 1D array instead of a 2D array were not sound as
written, in part because later stages in the process (after
edge-detection) need to now dispatch on whether they happen to be
handed a 1D ParallelArray or a 2D one, so that they can call the
appropriate get() method.  (Also, I still need to confirm absolutely
that switching to a 1D ParallelArray was a critically required change
or not.)

Also, since we no longer universally construct a 2D array, I need to
carry around the intended width and height explicitly again (or pass
it as separate parameters, like the original benchmark did).

With these fixes in place, it looks like the parallel execution may
produce the same result as the sequential one (a better place to start
from when evaluating performance issues).