fix bottle-neck: ParArr construction slow with fill fcn r.get(i,j).
This replaces:

    function (i,j) r.get(i,j)

with

    function (i,j) b[i+w*j]

where b = r.backingArray (i.e. manually lifted constant from loop) and
r.get method is defined by:

    WrapArray.prototype.get = function get(x,y,j) {
      if (j === undefined) j = 0;
      return this.backingArray[(y*this.width+x)*this.payload+j];
    };

This transformation yields a 10x speedup.  (I have not yet checked
which parts of our pipeline this speedup is helping; it could be
isolated to ParallelArray, or it could be the whole SpiderMonkey
pipeline.)