Bottle-neck fix for seq: Replace r.get(i,j) with b[i+w*j].
I recently put in a fix for a bottleneck in ParallelArray
construction.  However, if I am going to put abstraction-breaking
hacks into ParallelArray, then I should put not penalize the
sequential version of the benchmark by leaving out analogous fixes
there.

Since I only benchmark construction and edge-deteciton (and leave the
latter stages in the pipeline as sequential-only for now), it was only
necessary to apply this policy to the sequential edge detection code.